requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
blinker==1.8.2
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7.1
Django==4.2.19
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.12
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.25.0
MarkupSafe==2.1.5
msgpack==1.1.0
packaging==24.2
pillow==10.4.0
psutil==7.0.0
psycopg2==2.9.3
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.4
roundrobin==0.0.4
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
